refactor import of utils function
