+++ scripts/generate_minknow_samplesheet.py
@@ -8,15 +8,17 @@ from argparse import ArgumentParser
 from datetime import datetime as dt
 
 import pandas as pd
-import psycopg2
-import yaml
 from genologics.config import BASEURI, PASSWORD, USERNAME
 from genologics.entities import Artifact, Process
 from genologics.lims import Lims
 from tabulate import tabulate
 
 from data.ONT_barcodes import format_well, ont_label2dict
-from scilifelab_epps.epp import traceback_to_step, upload_file
+from scilifelab_epps.epp import (
+    get_pool_sample_label_mapping,
+    traceback_to_step,
+    upload_file,
+)
 from scilifelab_epps.utils.udf_tools import fetch
 from scilifelab_epps.wrapper import epp_decorator
 
@@ -189,64 +191,6 @@ def get_ont_library_contents(
         )
 
     return df
-
-
-def get_pool_sample_label_mapping(pool: Artifact) -> dict[str, str]:
-    with open("/opt/gls/clarity/users/glsai/config/genosqlrc.yaml") as f:
-        config = yaml.safe_load(f)
-
-    # Setup DB connection
-    connection = psycopg2.connect(
-        user=config["username"],
-        host=config["url"],
-        database=config["db"],
-        password=config["password"],
-    )
-    cursor = connection.cursor()
-
-    # Find all reagent labels linked to 'analyte' type artifacts matching the given name
-    query = """
-        select
-            distinct( rl.name )
-        from
-            reagentlabel            rl,
-            artifact                art,
-            artifact_label_map      alm
-        where
-            rl.labelid              = alm.labelid
-            and art.artifactid      = alm.artifactid
-            and art.artifacttypeid  = 2
-            and art.name            = '{}';
-    """
-
-    errors = False
-    sample2label = {}
-    for sample in pool.samples:
-        try:
-            cursor.execute(query.format(sample.name))
-            query_results = cursor.fetchall()
-
-            assert len(query_results) != 0, (
-                f"No reagent labels found for sample '{sample.name}'."
-            )
-            assert len(query_results) == 1, (
-                f"Multiple reagent labels found for sample '{sample.name}'."
-            )
-
-            label = query_results[0][0]
-            sample2label[sample.name] = label
-        except AssertionError as e:
-            logging.error(str(e), exc_info=True)
-            logging.warning(f"Skipping sample '{sample.name}' due to error.")
-            errors = True
-            continue
-
-    if errors:
-        raise AssertionError(
-            "Errors occurred when linking samples and indices. Please report this error."
-        )
-    else:
-        return sample2label
 
 
 def get_kit_string(process: Process) -> str:
